+++ scripts/build_central_heating_temperature_profiles/run.py
@@ -192,7 +192,7 @@ if __name__ == "__main__":
         from scripts._helpers import mock_snakemake
 
         snakemake = mock_snakemake(
-            "build_cop_profiles",
+            "build_central_heating_temperature_profiles",
             clusters=48,
             planning_horizons="2050",
         )